chore(ingestion): move jfr parser to its own package.
This should prevent logrus to become a dependency of the go profiler.
